refactor(utilization): extract UtilizationDao from FacilityDao
